1. Moved validTokenChars from HttpHeaderValidationUtil to HttpUtil 2. Added HTTP Method validation in HttpMethod 3. Reverted code from the previous approach
