+++ codec-http/src/main/java/io/netty/handler/codec/http/HttpHeaderValidationUtil.java
@@ -161,138 +161,4 @@ public final class HttpHeaderValidationUtil {
         }
         return -1;
     }
-
-    /**
-     * Validate a <a href="https://tools.ietf.org/html/rfc7230#section-3.2.6">token</a> contains only allowed
-     * characters.
-     * <p>
-     * The <a href="https://tools.ietf.org/html/rfc2616#section-2.2">token</a> format is used for variety of HTTP
-     * components, like  <a href="https://tools.ietf.org/html/rfc6265#section-4.1.1">cookie-name</a>,
-     * <a href="https://tools.ietf.org/html/rfc7230#section-3.2.6">field-name</a> of a
-     * <a href="https://tools.ietf.org/html/rfc7230#section-3.2">header-field</a>, or
-     * <a href="https://tools.ietf.org/html/rfc7231#section-4">request method</a>.
-     *
-     * @param token the token to validate.
-     * @return the index of the first invalid token character found, or {@code -1} if there are none.
-     */
-    public static int validateToken(CharSequence token) {
-        if (token instanceof AsciiString) {
-            return validateAsciiStringToken((AsciiString) token);
-        }
-        return validateCharSequenceToken(token);
-    }
-
-    /**
-     * Validate that an {@link AsciiString} contain onlu valid
-     * <a href="https://tools.ietf.org/html/rfc7230#section-3.2.6">token</a> characters.
-     *
-     * @param token the ascii string to validate.
-     */
-    private static int validateAsciiStringToken(AsciiString token) {
-        byte[] array = token.array();
-        for (int i = token.arrayOffset(), len = token.arrayOffset() + token.length(); i < len; i++) {
-            if (!BitSet128.contains(array[i], TOKEN_CHARS_HIGH, TOKEN_CHARS_LOW)) {
-                return i - token.arrayOffset();
-            }
-        }
-        return -1;
-    }
-
-    /**
-     * Validate that a {@link CharSequence} contain onlu valid
-     * <a href="https://tools.ietf.org/html/rfc7230#section-3.2.6">token</a> characters.
-     *
-     * @param token the character sequence to validate.
-     */
-    private static int validateCharSequenceToken(CharSequence token) {
-        for (int i = 0, len = token.length(); i < len; i++) {
-            byte value = (byte) token.charAt(i);
-            if (!BitSet128.contains(value, TOKEN_CHARS_HIGH, TOKEN_CHARS_LOW)) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
-    private static final long TOKEN_CHARS_HIGH;
-    private static final long TOKEN_CHARS_LOW;
-    static {
-        // HEADER
-        // header-field   = field-name ":" OWS field-value OWS
-        //
-        // field-name     = token
-        // token          = 1*tchar
-        //
-        // tchar          = "!" / "#" / "$" / "%" / "&" / "'" / "*"
-        //                    / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~"
-        //                    / DIGIT / ALPHA
-        //                    ; any VCHAR, except delimiters.
-        //  Delimiters are chosen
-        //   from the set of US-ASCII visual characters not allowed in a token
-        //   (DQUOTE and "(),/:;<=>?@[\]{}")
-        //
-        // COOKIE
-        // cookie-pair       = cookie-name "=" cookie-value
-        // cookie-name       = token
-        // token          = 1*<any CHAR except CTLs or separators>
-        // CTL = <any US-ASCII control character
-        //       (octets 0 - 31) and DEL (127)>
-        // separators     = "(" | ")" | "<" | ">" | "@"
-        //                      | "," | ";" | ":" | "\" | <">
-        //                      | "/" | "[" | "]" | "?" | "="
-        //                      | "{" | "}" | SP | HT
-        //
-        // field-name's token is equivalent to cookie-name's token, we can reuse the tchar mask for both:
-        BitSet128 tokenChars = new BitSet128()
-                .range('0', '9').range('a', 'z').range('A', 'Z') // Alphanumeric.
-                .bits('-', '.', '_', '~') // Unreserved characters.
-                .bits('!', '#', '$', '%', '&', '\'', '*', '+', '^', '`', '|'); // Token special characters.
-        TOKEN_CHARS_HIGH = tokenChars.high();
-        TOKEN_CHARS_LOW = tokenChars.low();
-    }
-
-    private static final class BitSet128 {
-        private long high;
-        private long low;
-
-        BitSet128 range(char fromInc, char toInc) {
-            for (int bit = fromInc; bit <= toInc; bit++) {
-                if (bit < 64) {
-                    low |= 1L << bit;
-                } else {
-                    high |= 1L << bit - 64;
-                }
-            }
-            return this;
-        }
-
-        BitSet128 bits(char... bits) {
-            for (char bit : bits) {
-                if (bit < 64) {
-                    low |= 1L << bit;
-                } else {
-                    high |= 1L << bit - 64;
-                }
-            }
-            return this;
-        }
-
-        long high() {
-            return high;
-        }
-
-        long low() {
-            return low;
-        }
-
-        static boolean contains(byte bit, long high, long low) {
-            if (bit < 0) {
-                return false;
-            }
-            if (bit < 64) {
-                return 0 != (low & 1L << bit);
-            }
-            return 0 != (high & 1L << bit - 64);
-        }
-    }
 }
